Make option status section component individually from share component
