+++ src/components/share/Share.scss
@@ -13,69 +13,5 @@
       width: 100%;
       margin: 20px 0 10px 0;
     }
-
-    .share-status-options-menu {
-      background-color: rgb(242, 240, 240);
-      padding: 7px;
-      border-radius: 5px;
-      margin-bottom: 10px;
-      display: flex;
-      flex-direction: column;
-      width: 100%;
-      height: 100%;
-      gap: 10px;
-      
-      @media only screen and (min-width: 1024px) {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        justify-content: space-between;
-        height: 80px;
-        gap: 6px;
-      }
-
-      .share-status-option-item {
-        flex: 4;
-        text-align: center;
-        background-color: rgb(234, 229, 229);
-        padding: 4px;
-        border-radius: 5px;
-        cursor: pointer;
-        width: 100%;
-        height: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-
-        &:hover {
-          background-color: rgb(163, 170, 238);
-          color: white;
-        }
-
-        &.active {
-          background-color: rgb(119, 131, 238);
-          color: white;
-        }
-
-        .share-status-name {
-          font-size: 11px;
-          margin-bottom: 4px;
-          
-          @media only screen and (min-width: 1024px) {
-            font-size: 12px;
-            margin-bottom: 6px;
-          }
-        }
-
-        .share-status-description {
-          font-size: 10px;
-
-          @media only screen and (min-width: 1024px) {
-            font-size: 10px;
-          }
-        }
-      }
-    }
   }
 }
